requirements.txt
# Core dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Model backends
openai>=1.0.0
anthropic>=0.8.1
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # For 8-bit quantization

# Web interface
streamlit>=1.29.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.1.0
